chore(refactor): move generic popup out of feel-popup feature
