--- src/features/feel-popup/index.js
@@ -1,8 +1,0 @@
-import { FeelPopup } from './FeelPopup';
-import { FeelPopupRenderer } from './FeelPopupRenderer';
-
-export default {
-  __init__: [ 'feelPopup', 'feelPopupRenderer' ],
-  feelPopup: [ 'type', FeelPopup ],
-  feelPopupRenderer: [ 'type', FeelPopupRenderer ],
-};
