+++ src/features/popup/index.js
@@ -1,0 +1,8 @@
+import { Popup } from './Popup';
+import { PopupRenderer } from './PopupRenderer';
+
+export default {
+  __init__: [ 'feelPopup', 'feelPopupRenderer' ],
+  feelPopup: [ 'type', Popup ],
+  feelPopupRenderer: [ 'type', PopupRenderer ],
+};
